[core][video] Move useReleasingSharedObject to expo-modules-core
