+++ packages/expo-video/src/VideoView.tsx
@@ -1,13 +1,5 @@
-import { SharedObject } from 'expo-modules-core';
-import {
-  ReactNode,
-  PureComponent,
-  DependencyList,
-  createRef,
-  useRef,
-  useMemo,
-  useEffect,
-} from 'react';
+import { useReleasingSharedObject } from 'expo-modules-core';
+import { ReactNode, PureComponent, createRef } from 'react';
 
 import NativeVideoModule from './NativeVideoModule';
 import NativeVideoView from './NativeVideoView';
@@ -97,52 +89,4 @@ function getPlayerId(player: number | VideoPlayer): number | null {
     return player;
   }
   return null;
-}
-
-/**
- * Returns a shared object, which is automatically cleaned up when the component is unmounted.
- */
-function useReleasingSharedObject<T extends SharedObject>(
-  factory: () => T,
-  dependencies: DependencyList
-): T {
-  const objectRef = useRef<T | null>(null);
-  const isFastRefresh = useRef(false);
-  const previousDependencies = useRef<DependencyList>(dependencies);
-
-  if (objectRef.current == null) {
-    objectRef.current = factory();
-  }
-
-  const object = useMemo(() => {
-    let newObject = objectRef.current;
-    const dependenciesAreEqual =
-      previousDependencies.current?.length === dependencies.length &&
-      dependencies.every((value, index) => value === previousDependencies.current[index]);
-
-    // If the dependencies have changed, release the previous object and create a new one, otherwise this has been called
-    // because of a fast refresh, and we don't want to release the object.
-    if (!newObject || !dependenciesAreEqual) {
-      objectRef.current?.release();
-      newObject = factory();
-      objectRef.current = newObject;
-      previousDependencies.current = dependencies;
-    } else {
-      isFastRefresh.current = true;
-    }
-    return newObject;
-  }, dependencies);
-
-  useEffect(() => {
-    isFastRefresh.current = false;
-
-    return () => {
-      // This will be called on every fast refresh and on unmount, but we only want to release the object on unmount.
-      if (!isFastRefresh.current && objectRef.current) {
-        objectRef.current.release();
-      }
-    };
-  }, []);
-
-  return object;
 }
